+++ cjsh-isocline/src/core/isocline_options.c
@@ -615,34 +615,34 @@ ic_public size_t ic_get_completion_menu_max_line_count(void) {
 ic_public size_t ic_set_history_menu_max_line_count(size_t line_count) {
     ic_env_t* env = ic_get_env();
     return ic_set_menu_line_limit(env == NULL ? NULL : &env->history_menu_max_line_count,
-                                  line_count, 30);
+                                  line_count, 15);
 }
 
 ic_public size_t ic_get_history_menu_max_line_count(void) {
     ic_env_t* env = ic_get_env();
-    return (env == NULL ? 30 : env->history_menu_max_line_count);
+    return (env == NULL ? 15 : env->history_menu_max_line_count);
 }
 
 ic_public size_t ic_set_command_palette_max_line_count(size_t line_count) {
     ic_env_t* env = ic_get_env();
     return ic_set_menu_line_limit(env == NULL ? NULL : &env->command_palette_max_line_count,
-                                  line_count, 30);
+                                  line_count, 15);
 }
 
 ic_public size_t ic_get_command_palette_max_line_count(void) {
     ic_env_t* env = ic_get_env();
-    return (env == NULL ? 30 : env->command_palette_max_line_count);
+    return (env == NULL ? 15 : env->command_palette_max_line_count);
 }
 
 ic_public size_t ic_set_custom_menu_max_line_count(size_t line_count) {
     ic_env_t* env = ic_get_env();
     return ic_set_menu_line_limit(env == NULL ? NULL : &env->custom_menu_max_line_count,
-                                  line_count, 30);
+                                  line_count, 15);
 }
 
 ic_public size_t ic_get_custom_menu_max_line_count(void) {
     ic_env_t* env = ic_get_env();
-    return (env == NULL ? 30 : env->custom_menu_max_line_count);
+    return (env == NULL ? 15 : env->custom_menu_max_line_count);
 }
 
 ic_public bool ic_enable_line_numbers(bool enable) {